feat(migration): change work hour columns from decimal to integer in presales_work_hours table

+++ AspireApp1.DbApi/Models/PreSalesWorkHour.cs
@@ -7,8 +7,8 @@ public class PreSalesWorkHour
     public string Title { get; set; } = string.Empty;
     public string? Description { get; set; }
     public int NumberOfPeople { get; set; }
-    public decimal WorkingHours { get; set; }
-    public decimal HourlyWage { get; set; }
+    public int WorkingHours { get; set; }
+    public int HourlyWage { get; set; }
     public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
     public DateTime? UpdatedAt { get; set; }
 
@@ -16,5 +16,5 @@ public class PreSalesWorkHour
     public PreSalesProposal PreSalesProposal { get; set; } = null!;
 
     // Calculated property
-    public decimal TotalCost => HourlyWage * NumberOfPeople * WorkingHours;
+    public int TotalCost => HourlyWage * NumberOfPeople * WorkingHours;
 }
